cleaning code

+++ backend/routes/api/exercises.js
@@ -52,28 +52,6 @@ const validateCreate = [
         .notEmpty()
         .isLength({ min: 1, max: 30 })
         .withMessage('Please provide a valid Name.'),
-    // check('distance')
-    //     .isInt({ min: 0, max: 10000 })
-    //     .withMessage('Please provide a valid distance [0-10000].'),
-    // check('sets')
-    //     .isInt({ min: 1, max: 1000 })
-    //     .withMessage('Please provide a valid set [0-1000].'),
-    // check('reps')
-    //     .isInt({ min: 1, max: 1000 })
-    //     .withMessage('Please provide valid reps [0-1000].'),
-    // check('rest')
-    //     .isInt({ min: 0, max: 86400 })
-    //     .withMessage('Please provide valid rest [0-86400].'),
-    // check('weight')
-    //     .isInt({ min: 1, max: 1500 })
-    //     .withMessage('Please provide valid reps [0-1500].'),
-    // check('notes')
-    //     .isLength({ max: 30 })
-    //     .withMessage('Please provide a note with less than 500 characters.'),
-    // check('completion_time')
-    //     .isInt({ min: 1, max: 86400 })
-    //     .toInt()
-    //     .withMessage('Workout time should be less than a whole day.'),
     handleValidationErrors
 ];
 
@@ -169,10 +147,10 @@ router.put(
     secondValidate,
     asyncHandler(async (req, res) => {
         const exercise = await Exercise.findByPk(req.params.id)
-        // if (req.user.id !== exercise.user_id) {
-        //     throw new Error('Access Denied. Your IP will be blocked and reported for suspicious activity. \n (Not really because this is a demo project, but it definitely will for the completed version.)');
-        //     return
-        // }
+        if (req.user.id !== exercise.user_id) {
+            throw new Error('Access Denied. Your IP will be blocked and reported for suspicious activity. \n (Not really because this is a demo project, but it definitely will for the completed version.)');
+            return
+        }
         exercise.name = req.body.name;
         exercise.distance = req.body.distance;
         exercise.sets = req.body.sets;
@@ -194,10 +172,10 @@ router.delete(
     restoreUser,
     asyncHandler(async function (req, res) {
         const exercise = await Exercise.findByPk(req.params.id);
-        // if (req.user.id !== Exercise.user_id) {
-        //     throw new Error('Access Denied. Your IP will be blocked and reported for suspicious activity. \n (Not really because this is a demo project, but it definitely will for the completed version.)');
-        //     return
-        // }
+        if (req.user.id !== Exercise.user_id) {
+            throw new Error('Access Denied. Your IP will be blocked and reported for suspicious activity. \n (Not really because this is a demo project, but it definitely will for the completed version.)');
+            return
+        }
         await Exercise.destroy({ where: { id: exercise.id } })
         return res.json(exercise)
     })
